requirements.txt
Mastodon.py==1.8.1
google-generativeai==0.3.1
python-dotenv==1.0.0
nltk==3.8.1
pillow==10.0.0
aiohttp==3.9.1